Renamed Option parser class to config factory and moved this and config class to Config folder
